+++ src/effects/events/SgEmitter.js
