+++ src/effects/events/SgEmitter.spec.js
