+++ src/effects/events/spawn.spec.js
