+++ src/effects/events/take.js
@@ -1,4 +1,4 @@
-import createEffect from './createEffect';
+import createEffect from '../createEffect';
 
 export const handleTakeEffect = ([type], emitter, id) =>
     emitter.take(id, type);
