+++ src/effects/events/takeEvery.spec.js
