+++ src/effects/events/takeLatest.spec.js
