+++ src/effects/index.js
@@ -1,15 +1,9 @@
+import * as eventsEffects from './events';
+
+export const events = eventsEffects;
 export { default as call } from './call';
-export { default as cancel } from './cancel';
+export { default as thunk } from './thunk';
 export { default as co } from './co';
 export { default as cps } from './cps';
-export { default as createEffect } from './createEffect';
-export { default as fork } from './fork';
-export { default as put } from './put';
-export { default as spawn } from './spawn';
-export { default as take } from './take';
-export { default as takeEvery } from './takeEvery';
-export { default as thunk } from './thunk';
-export { default as takeLatest } from './takeLatest';
 export { default as delay } from './delay';
-export { default as race } from './race';
-export { default as join } from './join';
+export { default as createEffect } from './createEffect';
