+++ src/utils/newTask.js
@@ -1,10 +1,9 @@
 import uuid from 'uuid';
-import SgEmitter from './SgEmitter';
+import SgEmitter from '../effects/events/SgEmitter';
 import isGenerator from './isGenerator';
 import deferred from './deferred';
 import effectHandler from './effectHandler';
 import sagaIterator from './sagaIterator';
-
 
 export default function newTask(generator, emitter, parentId = null) {
     const id = uuid();
